refactor: Servidores movidos a su propia carpeta
